backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
gunicorn
whitenoise
dj-database-url
psycopg2-binary
openai
